+++ src/playlistplayer.cpp
@@ -103,10 +103,10 @@ void PlaylistPlayer::libvlc_callback(const libvlc_event_t *event, void *data)
         currentPlayback = core->_playlist->at(core->_currentIndex);
 
         connect(currentPlayback->mediaSettings(), SIGNAL(ratioChanged(Ratio)), core->_mediaPlayer, SLOT(setCurrentRatio(Ratio)));
-        connect(currentPlayback->mediaSettings(), SIGNAL(gammaChanged(int)), core->_mediaPlayer, SLOT(setCurrentGamma(int)));
-        connect(currentPlayback->mediaSettings(), SIGNAL(contrastChanged(int)), core->_mediaPlayer, SLOT(setCurrentContrast(int)));
-        connect(currentPlayback->mediaSettings(), SIGNAL(brightnessChanged(int)), core->_mediaPlayer, SLOT(setCurrentBrightness(int)));
-        connect(currentPlayback->mediaSettings(), SIGNAL(saturationChanged(int)), core->_mediaPlayer, SLOT(setCurrentSaturation(int)));
+        connect(currentPlayback->mediaSettings(), SIGNAL(gammaChanged(float)), core->_mediaPlayer, SLOT(setCurrentGamma(float)));
+        connect(currentPlayback->mediaSettings(), SIGNAL(contrastChanged(float)), core->_mediaPlayer, SLOT(setCurrentContrast(float)));
+        connect(currentPlayback->mediaSettings(), SIGNAL(brightnessChanged(float)), core->_mediaPlayer, SLOT(setCurrentBrightness(float)));
+        connect(currentPlayback->mediaSettings(), SIGNAL(saturationChanged(float)), core->_mediaPlayer, SLOT(setCurrentSaturation(float)));
         connect(currentPlayback->mediaSettings(), SIGNAL(hueChanged(int)), core->_mediaPlayer, SLOT(setCurrentHue(int)));
         connect(currentPlayback->mediaSettings(), SIGNAL(deinterlacingChanged(Deinterlacing)), core->_mediaPlayer, SLOT(setCurrentDeinterlacing(Deinterlacing)));
         connect(currentPlayback->mediaSettings(), SIGNAL(subtitlesSyncChanged(double)), core->_mediaPlayer, SLOT(setCurrentSubtitlesSync(double)));
